Poprawiono błąd krytyczny związany z poszukiwaniem drogi przez ewakuantów. Teraz szukają najtańszego wyjscia, a nie najbliższego. Poprawia to błąd z nie poruszaniem agentów w razie znajdowania najbliższego wyjscia w niedostepnej dla niego częsci pomieszczenia. Poprawia to teraz logike poruszania ewakuantów i ich reakcję na zagrożenie. Teraz do strategii AStarPathFinder jest przykazywana lista wyjsc i juz strategia decyduje o najblizszym wyjsciu

+++ app/src/main/java/evacuation/sim/agent/human/Panicked.java
@@ -40,7 +40,7 @@ public class Panicked extends Evacuee{
                         int randomIndex = (int)(Math.random() * validChoices.size());
                         Cell randomDestination = validChoices.get(randomIndex);
 
-                        getPlannedPath().add(board.getCell((int)getLogicalX(), (int)getLogicalY()));
+                        getPlannedPath().add(board.getCell(getLogicalX(), getLogicalY()));
                         getPlannedPath().add(randomDestination);
                     }
             }
